requirements.txt
openai>=1.79.0
pandas>=1.5.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0

#pip install -r requirements.txt
